wip: register allocation mostly works
fails when a variable declared inside loop

seems like I'll need PHI nodes at some point

+++ src/vm.rs
@@ -11,8 +11,8 @@ pub enum Op {
 	LessThanEq,
 	GreaterThanEq,
 	/// idx is zero-indexed
-	LoadVar { stack_idx: u32 },
-	SetVar { stack_idx: u32 },
+	LoadVar { stack_idx: usize },
+	SetVar { stack_idx: usize },
 	JumpLabel { label_id: usize },
 	Jump { label_id: usize },
 	JumpIfZero { label_id: usize },
